renamed bug file
